pyproject.toml
[tool.poetry]
name = "test-pylsar-msk"
version = "0.1.0"
description = ""
authors = ["polenom <vitalimitsevich@yandex.by>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.4"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.5"
pytest = "^7.2.0"
pytest-django = "^4.5.2"
pillow = "^9.3.0"
requests = "^2.28.1"
django-filter = "^22.1"
drf-spectacular = "^0.25.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
